requirements.txt
autopep8==1.5.3
Babel==2.8.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
crispy-forms-materialize==0.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap-modal-forms==1.5.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-embed-video==1.3.2
django-floppyforms==1.9.0
django-froala-editor==3.1.0
django-imagekit==4.0.2
django-materialize==1.0.1a2
django-widget-tweaks==1.4.8
idna==2.9
mysqlclient==1.4.2.post1
oauthlib==3.1.0
pep8==1.7.1
pilkit==2.0
pycodestyle==2.6.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
